> - created/ accordion and tabs components > - coments/ useSearchParams in modalPortal and headerListIcons > - created/ file components put accordions > - created/ a11Tabs, accordions-panel and tab-panel utils - deleted/ style in account-page
