+++ src/components/Layout/header/header-list-icons/headerListIcons.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable prettier/prettier */
-import { Link, useSearchParams } from 'react-router-dom'
+import { Link } from 'react-router-dom'
 import { useSelector } from 'react-redux'
 import { Fragment, useEffect, useRef, useState } from 'react'
 import ModalPortal from '../../../modal-portal/ModalPortal'
@@ -18,13 +18,13 @@ const HeaderListIcons = () => {
   const inputRef = useRef<HTMLInputElement>(null)
   const isLoggedIn = useSelector((state: RootState) => state.auth.isLoggedIn)
 
-  const [searchParams] = useSearchParams()
+  // const [searchParams] = useSearchParams()
 
-  const searchToken = Object.fromEntries(searchParams.entries())
+  // const searchToken = Object.fromEntries(searchParams.entries())
 
-  useEffect(() => {
-    Object.keys(searchToken).length > 0 && setIsModalOpen(true)
-  }, [searchToken])
+  // useEffect(() => {
+  //   Object.keys(searchToken).length > 0 && setIsModalOpen(true)
+  // }, [searchToken])
 
   const handleClick = (event: MouseEvent) => {
     const target = event.target as HTMLElement
@@ -101,12 +101,8 @@ const HeaderListIcons = () => {
           </Link>
         </li>
       </ul>
-      <ModalPortal
-        isModalOpen={isModalOpen}
-        setIsModalOpen={setIsModalOpen}
-        searchToken={searchToken}
-      >
-        <Auth searchToken={searchToken} setIsModalOpen={setIsModalOpen} />
+      <ModalPortal isModalOpen={isModalOpen} setIsModalOpen={setIsModalOpen}>
+        <Auth setIsModalOpen={setIsModalOpen} />
       </ModalPortal>
     </Fragment>
   )
