+++ src/components/Modal-custom-btn/ModalCustomBtnEmail.tsx
@@ -59,7 +59,7 @@ export const EmailConfirmationModal = ({
 
   return (
     <Fragment>
-      {!is_active ? null : (
+      {is_active ? null : (
         <Link
           to="/account"
           onClick={() => {
