--- src/pages/account-page/style.ts
@@ -1,32 +1,0 @@
-export const stWavePink = {
-  backgroundImage: 'url("/src/icons/accountWave.svg")',
-  width: '100%',
-  backgroundRepeat: 'no-repeat',
-  backgroundSize: '100%',
-  position: 'relative',
-  bottom: 1
-}
-
-export const stTabsNav = {
-  width: '100%',
-  padding: '10px',
-  borderLeft: '1px solid',
-  borderTop: '1px solid',
-  borderBottom: '1px solid',
-  borderColor: 'peach.darker',
-  borderRadius: '30px 0 0 30px'
-}
-export const stTabsBottomBox = {
-  padding: '10px',
-  borderRight: '1px solid',
-  borderTop: '1px solid',
-  borderBottom: '1px solid',
-  borderColor: 'peach.darker',
-  borderRadius: ' 0  30px 30px 0'
-}
-
-export const stContainerTabPanel = {
-  backgroundColor: 'pink.lighter',
-  color: 'secondary.darker',
-  fontFamily: 'Comfortaa'
-}
