+++ src/utils/a11y-tabs/a11yTabs.ts
@@ -1,0 +1,4 @@
+export const a11yProps = (index: number) => ({
+  'id': `simple-tab-${index}`,
+  'aria-controls': `simple-tabpanel-${index}`
+})
